+++ src/containerapp/azext_containerapp/containerapp_decorator.py
@@ -1409,9 +1409,6 @@ class ContainerAppPreviewUpdateDecorator(ContainerAppUpdateDecorator):
     def set_argument_registry_identity(self, registry_identity):
         self.set_param("registry_identity", registry_identity)
 
-    def get_argument_kind(self):
-        return self.get_param("kind")
-
     def validate_arguments(self):
         super().validate_arguments()
         if self.get_argument_service_bindings() and len(self.get_argument_service_bindings()) > 1 and self.get_argument_customized_keys():
@@ -1534,9 +1531,6 @@ class ContainerAppPreviewUpdateDecorator(ContainerAppUpdateDecorator):
                     subscription_id = get_subscription_id(self.cmd.cli_ctx)
                     identity = _ensure_identity_resource_id(subscription_id, self.get_argument_resource_group_name(), identity)
                 self.new_containerapp["properties"]["template"]["scale"]["rules"][0]["custom"]["identity"] = identity
-
-        if self.get_argument_kind() is not None:
-            safe_set(self.new_containerapp, "kind", value=self.get_argument_kind())
 
     def set_up_source(self):
         from ._up_utils import (_validate_source_artifact_args)